COMP: Remove comparison of integers of different signs: 'int' and 'size_t'
Size of memory allocation bytes should be size_t.

nifti_clib/nifti2/nifti2_io.c:4404:19: warning: comparison of integers of different signs: 'int' and 'size_t' (aka 'unsigned long') [-Wsign-compare]
   for ( c = 0; c < nbytes; c++ )
              ~ ^ ~~~~~~
nifti_clib/nifti2/nifti2_io.c:6431:72: warning: format specifies type 'int' but the argument has type 'size_t' (aka 'unsigned long') [-Wformat]
   fprintf(stderr,"** %s: nbytes=%d, too small for test", fname, nbytes);
                                 ~~                              ^~~~~~
                                 %zu

+++ niftilib/nifti1_io.c
@@ -435,7 +435,7 @@ static int   is_uppercase      (const char * str);
 static int   make_lowercase    (char * str);
 static int   make_uppercase    (char * str);
 static int   need_nhdr_swap    (short dim0, int hdrsize);
-static int   print_hex_vals    (const char * data, int nbytes, FILE * fp);
+static int   print_hex_vals    (const char * data, size_t nbytes, FILE * fp);
 static int   unescape_string   (char *str);  /* string utility functions */
 static char *escapize_string   (const char *str);
 
@@ -3501,9 +3501,9 @@ int is_nifti_file( const char *hname )
    return -1 ;                          /* not good */
 }
 
-static int print_hex_vals( const char * data, int nbytes, FILE * fp )
+static int print_hex_vals( const char * data, size_t nbytes, FILE * fp )
 {
-   int c;
+   size_t c;
 
    if ( !data || nbytes < 1 || !fp ) return -1;
